# packages/template/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "operationId": "Customers_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "summary": "Create a customer",
        "tags": [
          "customers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      },
      "get": {
        "operationId": "Customers_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerList"
                }
              }
            }
          }
        },
        "summary": "List customers",
        "tags": [
          "customers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "operationId": "Customers_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "summary": "Get a customer",
        "tags": [
          "customers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      },
      "patch": {
        "operationId": "Customers_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "summary": "Update a customer",
        "tags": [
          "customers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      },
      "delete": {
        "operationId": "Customers_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a customer",
        "tags": [
          "customers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      }
    },
    "/api/v1/plans": {
      "post": {
        "operationId": "Plans_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          }
        },
        "summary": "Create a plan (Stripe product + price)",
        "tags": [
          "plans"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      },
      "get": {
        "operationId": "Plans_list",
        "parameters": [
          {
            "name": "activeOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  }
                }
              }
            }
          }
        },
        "summary": "List plans",
        "tags": [
          "plans"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      }
    },
    "/api/v1/plans/{id}": {
      "get": {
        "operationId": "Plans_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          }
        },
        "summary": "Get a plan",
        "tags": [
          "plans"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      }
    },
    "/api/v1/plans/{id}/archive": {
      "post": {
        "operationId": "Plans_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          }
        },
        "summary": "Archive a plan (prices are immutable in Stripe)",
        "tags": [
          "plans"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      }
    },
...
    "/api/v1/subscriptions": {
      "post": {
        "operationId": "Subscriptions_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSubscriptionResult"
                }
              }
            }
          }
        },
        "summary": "Create a subscription (first charge confirmed by frontend)",
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      },
      "get": {
        "operationId": "Subscriptions_list",
        "parameters": [
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionList"
                }
              }
            }
          }
        },
        "summary": "List subscriptions, optionally by customer",
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      }
    },
    "/api/v1/subscriptions/{id}": {
      "get": {
        "operationId": "Subscriptions_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        },
        "summary": "Get a subscription",
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      },
      "delete": {
        "operationId": "Subscriptions_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        },
        "summary": "Cancel a subscription immediately",
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Stripe Payment Service",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "phone": {
            "type": "string",
            "example": "+15555555555"
          }
        },
        "required": [
          "email"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "stripeCustomerId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stripeCustomerId",
          "email",
          "name",
          "phone",
          "createdAt",
          "updatedAt"
        ]
      },
      "CustomerList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Customer"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "new@example.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "phone": {
            "type": "string",
            "example": "+15555555555"
          }
        }
      },
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Pro Monthly"
          },
          "amount": {
            "type": "number",
            "minimum": 1,
            "example": 1900,
            "description": "Amount in smallest currency unit (cents)"
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "example": "usd"
          },
          "interval": {
            "enum": [
              "month",
              "year"
            ],
            "type": "string",
            "example": "month"
          }
        },
        "required": [
          "name",
          "amount",
          "currency",
          "interval"
        ]
      },
      "Plan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "stripeProductId": {
            "type": "string"
          },
          "stripePriceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "description": "Smallest currency unit (cents) — never floating point"
          },
          "currency": {
            "type": "string"
          },
          "interval": {
            "enum": [
              "year",
              "month"
            ],
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stripeProductId",
          "stripePriceId",
          "name",
          "amount",
          "currency",
          "interval",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateSubscriptionDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid",
            "description": "Local customer id (UUID)"
          },
          "planId": {
            "type": "string",
            "format": "uuid",
            "description": "Local plan id (UUID)"
          },
          "trialPeriodDays": {
            "type": "number",
            "minimum": 1,
            "example": 14,
            "description": "Free trial length in days — Stripe schedules the first charge"
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Pass the same key on retries to guarantee the subscription is created at most once"
          }
        },
        "required": [
          "customerId",
          "planId"
        ]
      },
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "stripeSubscriptionId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "planId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "canceled",
              "incomplete",
              "incomplete_expired",
              "trialing",
              "past_due",
              "unpaid"
            ],
            "type": "string"
          },
          "currentPeriodStart": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "currentPeriodEnd": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "trialEnd": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stripeSubscriptionId",
          "customerId",
          "planId",
          "status",
          "currentPeriodStart",
          "currentPeriodEnd",
          "trialEnd",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateSubscriptionResult": {
        "type": "object",
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/Subscription"
          },
          "clientSecret": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "subscription",
          "clientSecret"
        ]
      },
      "SubscriptionList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Subscription"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
...
    }
  }
}